Normalize Github event context
Events coming in from Github have a special kind of "context" indicating
the type of event. This is subject to change and we don't really care
for the details -- all we want to know is the _kind_ of service. So omit
the Codeship part, and only store that it is a "ci" event.

This removes the need to extract a "ci" const in our client side code,
making it simpler. It might still be a good idea to extract such values
into constants, but it's not needed just yet.